requirements.txt
Flask==2.2.2
tensorflow>=2.13.0
firebase-admin==6.0.1
google-api-python-client==2.70.0
google-auth==2.16.0
PyPDF2==3.0.1
numpy==1.26.4
google-auth-oauthlib==0.4.6
Werkzeug==2.2.2
transformers==4.50.0
tf-keras==2.19.0